refactored to util-db subcommand instead specific binary
