+++ cmd/util-db/primer/run_primer.go
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with Aida. If not, see <http://www.gnu.org/licenses/>.
 
-package db
+package primer
 
 import (
 	"github.com/urfave/cli/v2"
